+++ spec/models/invoice_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Invoice, type: :model do
     it { is_expected.to have_db_column(:id).of_type(:uuid) }
     it { is_expected.to have_db_column(:community_id).of_type(:uuid) }
     it { is_expected.to have_db_column(:land_parcel_id).of_type(:uuid) }
-    it { is_expected.to have_db_column(:amount).of_type(:integer) }
+    it { is_expected.to have_db_column(:amount).of_type(:float) }
     it { is_expected.to have_db_column(:note).of_type(:string) }
     it { is_expected.to have_db_column(:description).of_type(:string) }
   end
