+++ spec/models/payment_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Payment, type: :model do
     it { is_expected.to have_db_column(:id).of_type(:uuid) }
     it { is_expected.to have_db_column(:user_id).of_type(:uuid) }
     it { is_expected.to have_db_column(:invoice_id).of_type(:uuid) }
-    it { is_expected.to have_db_column(:amount).of_type(:integer) }
+    it { is_expected.to have_db_column(:amount).of_type(:float) }
     it { is_expected.to have_db_column(:payment_status).of_type(:integer) }
     it { is_expected.to have_db_column(:payment_type).of_type(:string) }
   end
